feat(): support standalone micro-apps
