+++ .gitignore
@@ -1,4 +1,5 @@
 dist
+dist-*
 node_modules
 .coverage
 .vscode
