--- packages/brick-container/.gitignore
@@ -1,1 +1,0 @@
-/dev-settings.yaml
