+++ packages/brick-container/src/browse-happy.ejs
@@ -3,7 +3,7 @@
   <head>
     <meta charset="utf-8">
     <title><%= htmlWebpackPlugin.options.title %></title>
-    <link rel="shortcut icon" type="image/png" href="assets/favicon.png"/>
+    <link rel="shortcut icon" type="image/png" href="<%= htmlWebpackPlugin.options.faviconPath %>"/>
   </head>
   <body>
     <p><a href="https://browsehappy.com/">请升级您的浏览器。</a></p>
