+++ packages/brick-container/webpack/prod.js
@@ -1,0 +1,55 @@
+const path = require("path");
+const { CleanWebpackPlugin } = require("clean-webpack-plugin");
+const MiniCssExtractPlugin = require("mini-css-extract-plugin");
+const { lessReplacePlugin } = require("@next-core/less-plugin-css-variables");
+
+module.exports = ({ standalone } = {}) => ({
+  mode: "production",
+  output: {
+    filename: "[name].[contenthash].js",
+  },
+  devtool: "source-map",
+  module: {
+    rules: [
+      {
+        test: /\.less$/,
+        sideEffects: true,
+        use: [
+          MiniCssExtractPlugin.loader,
+          "css-loader",
+          {
+            loader: "less-loader",
+            options: {
+              lessOptions: {
+                sourceMap: true,
+                javascriptEnabled: true,
+                plugins: [lessReplacePlugin],
+              },
+            },
+          },
+        ],
+      },
+      {
+        test: /\.css$/,
+        // Ref https://github.com/webpack-contrib/mini-css-extract-plugin/issues/118
+        sideEffects: true,
+        use: [MiniCssExtractPlugin.loader, "css-loader"],
+      },
+    ],
+  },
+  plugins: [
+    new CleanWebpackPlugin({
+      cleanOnceBeforeBuildPatterns: [
+        standalone
+          ? path.resolve(__dirname, "../dist-standalone/**/*")
+          : "**/*",
+      ],
+    }),
+    new MiniCssExtractPlugin({
+      // Options similar to the same options in webpackOptions.output
+      // both options are optional
+      filename: "[name].[contenthash].css",
+      chunkFilename: "[id].css",
+    }),
+  ],
+});
